Endringer etter qa
