+++ melosys-eessi-app/src/main/java/no/nav/melosys/eessi/integration/sak/SakConsumer.java
@@ -1,6 +1,7 @@
 package no.nav.melosys.eessi.integration.sak;
 
 import lombok.extern.slf4j.Slf4j;
+import no.nav.melosys.eessi.config.MDCOperations;
 import no.nav.melosys.eessi.integration.RestConsumer;
 import no.nav.melosys.eessi.models.exception.IntegrationException;
 import org.springframework.http.HttpEntity;
@@ -12,8 +13,6 @@ import org.springframework.web.client.RestTemplate;
 @Slf4j
 public class SakConsumer implements RestConsumer {
 
-    private static final String X_CORRELATION_ID = "X-Correlation-ID";
-
     private final RestTemplate restTemplate;
 
     public SakConsumer(RestTemplate restTemplate) {
@@ -23,7 +22,7 @@ public class SakConsumer implements RestConsumer {
     public Sak getSak(String arkivsakId) {
 
         HttpHeaders headers = headers();
-        log.info("hentsak: correlationId: {}, sakId: {}", headers.get(X_CORRELATION_ID), arkivsakId);
+        log.info("hentsak: correlationId: {}, sakId: {}", headers.get(MDCOperations.X_CORRELATION_ID), arkivsakId);
 
         return exchange("/{arkivsakId}", HttpMethod.GET, new HttpEntity<>(headers), Sak.class, arkivsakId);
     }
