+++ melosys-eessi-app/src/test/resources/logback.xml
@@ -4,7 +4,7 @@
     <!-- encoders are assigned the type
          ch.qos.logback.classic.encoder.PatternLayoutEncoder by default -->
     <encoder>
-      <pattern>%d{HH:mm:ss.SSS} [%thread] %-5level %logger{36} - %msg %mdc%n</pattern>
+        <pattern>%d{HH:mm:ss.SSS} | %X{correlation-id} | %logger{10} | %highlight(%level) | %msg%n</pattern>
     </encoder>
   </appender>
