+++ melosys-eessi-test/src/test/resources/logback-test.xml
@@ -3,7 +3,7 @@
 
   <appender name="console" class="ch.qos.logback.core.ConsoleAppender">
     <encoder>
-      <pattern>%d{HH:mm:ss.SSS} %highlight(%-5level) %logger{10} [%X{sedId}] - %msg%n</pattern>
+        <pattern>%d{HH:mm:ss.SSS} | %X{correlation-id} | %logger{10} | %highlight(%level) | %msg%n</pattern>
     </encoder>
   </appender>
